Refactor (main_vision): standardize headers and comments, run at full resolution (drop 2x downscale), remove servo control and output vision over serial.
