+++ vision_orin/main_vision/orin_config.json
@@ -1,0 +1,39 @@
+{
+  "_config": "orin_config.json",
+  "_purpose": "HSV + geometry thresholds for the bean sorter vision pipeline (orin_vision.py / orin_vision_debug.py).",
+  "_note": "Coordinates and sizes are stored at full ZED resolution (1344x376); the scripts scale them down 2x at runtime.",
+  "_workflow": "Tune on a PC with orin_vision_debug.py until beans are picked cleanly, save, then copy this file onto the Orin.",
+  "_fields": {
+    "s_min": "min HSV saturation for a pixel to count as a bean candidate",
+    "v_dark": "pixels darker than this value also count as candidates",
+    "bg_h_lo/bg_h_hi": "hue range of the green background to subtract",
+    "bg_s_min/bg_v_min": "min saturation/value for the background subtraction",
+    "area_min/area_max": "accepted contour area range (full-res px)",
+    "rad_min/rad_max": "accepted enclosing-circle radius range (full-res px)",
+    "circ_min/circ_max": "accepted circularity range, as percent (0-100)",
+    "morph_k": "morphology kernel size (forced odd)",
+    "morph_iter": "morphology open/close iterations",
+    "det_cy_l/det_cy_r": "x-center of the detection zone for left/right halves",
+    "det_thick": "half-width of the detection zone",
+    "trigger_x/trig_y2": "y-band that a bean must cross to fire a hit"
+  },
+  "s_min": 145,
+  "v_dark": 62,
+  "bg_h_lo": 33,
+  "bg_h_hi": 104,
+  "bg_s_min": 0,
+  "bg_v_min": 0,
+  "area_min": 961,
+  "area_max": 5247,
+  "rad_min": 31,
+  "rad_max": 125,
+  "circ_min": 66,
+  "circ_max": 99,
+  "morph_k": 9,
+  "morph_iter": 3,
+  "det_cy_l": 286,
+  "det_cy_r": 379,
+  "det_thick": 97,
+  "trigger_x": 103,
+  "trig_y2": 376
+}
